Refactor edit page --> settings page
