+++ frontend/app/pages/settings/[id].vue
@@ -1,4 +1,4 @@
-﻿<script setup lang="ts">
+<script setup lang="ts">
 const _route = useRoute()
 </script>
